Extract optimizer configs into OptimizerConfig
There are a significant number of properties
related to optimizer rules in FeaturesConfig.
Extracting these to OptimizerConfig makes FeaturesConfig
smaller and provides a convenient grouping of optimizer
related flags in one place.
