+++ core/trino-main/src/main/java/io/trino/FeaturesConfig.java
@@ -21,7 +21,6 @@ import io.airlift.configuration.ConfigDescription;
 import io.airlift.configuration.DefunctConfig;
 import io.airlift.configuration.LegacyConfig;
 import io.airlift.units.DataSize;
-import io.airlift.units.Duration;
 import io.airlift.units.MaxDataSize;
 import io.trino.sql.analyzer.RegexLibrary;
 
@@ -36,10 +35,7 @@ import java.util.List;
 
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static io.airlift.units.DataSize.Unit.KILOBYTE;
-import static io.airlift.units.DataSize.Unit.MEGABYTE;
 import static io.trino.sql.analyzer.RegexLibrary.JONI;
-import static java.util.Objects.requireNonNull;
-import static java.util.concurrent.TimeUnit.MINUTES;
 
 @DefunctConfig({
         "analyzer.experimental-syntax-enabled",
@@ -72,39 +68,18 @@ public class FeaturesConfig
     static final String SPILL_ENABLED = "spill-enabled";
     public static final String SPILLER_SPILL_PATH = "spiller-spill-path";
 
-    private double cpuCostWeight = 75;
-    private double memoryCostWeight = 10;
-    private double networkCostWeight = 15;
-    private boolean distributedIndexJoinsEnabled;
-    private DataSize joinMaxBroadcastTableSize = DataSize.of(100, MEGABYTE);
-    private JoinDistributionType joinDistributionType = JoinDistributionType.AUTOMATIC;
-    private boolean colocatedJoinsEnabled;
     private boolean groupedExecutionEnabled;
     private boolean dynamicScheduleForGroupedExecution;
     private int concurrentLifespansPerTask;
-    private boolean spatialJoinsEnabled = true;
-    private JoinReorderingStrategy joinReorderingStrategy = JoinReorderingStrategy.AUTOMATIC;
-    private int maxReorderedJoins = 9;
     private boolean redistributeWrites = true;
-    private boolean usePreferredWritePartitioning = true;
-    private int preferredWritePartitioningMinNumberOfPartitions = 50;
     private boolean scaleWriters;
     private DataSize writerMinSize = DataSize.of(32, DataSize.Unit.MEGABYTE);
-    private boolean optimizeMetadataQueries;
-    private boolean optimizeHashGeneration = true;
-    private boolean enableIntermediateAggregations;
-    private boolean pushTableWriteThroughUnion = true;
     private DataIntegrityVerification exchangeDataIntegrityVerification = DataIntegrityVerification.ABORT;
     private boolean exchangeCompressionEnabled;
     private boolean legacyRowToJsonCast;
-    private boolean optimizeMixedDistinctAggregations;
-    private boolean forceSingleNodeOutput = true;
     private boolean pagesIndexEagerCompactionEnabled;
-    private boolean distributedSort = true;
     private boolean omitDateTimeTypePrecision;
     private int maxRecursionDepth = 10;
-
-    private boolean dictionaryAggregation;
 
     private int re2JDfaStatesLimit = Integer.MAX_VALUE;
     private int re2JDfaRetries = 5;
@@ -114,33 +89,11 @@ public class FeaturesConfig
     private List<Path> spillerSpillPaths = ImmutableList.of();
     private int spillerThreads = 4;
     private double spillMaxUsedSpaceThreshold = 0.9;
-    private boolean enableStatsCalculator = true;
-    private boolean statisticsPrecalculationForPushdownEnabled;
-    private boolean collectPlanStatisticsForAllQueries;
-    private boolean ignoreStatsCalculatorFailures = true;
-    private boolean defaultFilterFactorEnabled;
-    private boolean enableForcedExchangeBelowGroupId = true;
-    private boolean pushAggregationThroughOuterJoin = true;
-    private boolean pushPartialAggregationThoughJoin;
     private double memoryRevokingTarget = 0.5;
     private double memoryRevokingThreshold = 0.9;
     private boolean parseDecimalLiteralsAsDouble;
-    private boolean useMarkDistinct = true;
-    private boolean preferPartialAggregation = true;
-    private boolean optimizeTopNRanking = true;
     private boolean lateMaterializationEnabled;
-    private boolean skipRedundantSort = true;
-    private boolean complexExpressionPushdownEnabled = true;
-    private boolean predicatePushdownUseTableProperties = true;
-    private boolean ignoreDownstreamPreferences;
-    private boolean rewriteFilteringSemiJoinToInnerJoin = true;
-    private boolean optimizeDuplicateInsensitiveJoins = true;
-    private boolean useLegacyWindowFilterPushdown;
-    private boolean useTableScanNodePartitioning = true;
-    private double tableScanNodePartitioningMinBucketToTaskRatio = 0.5;
-    private boolean mergeProjectWithValues = true;
 
-    private Duration iterativeOptimizerTimeout = new Duration(3, MINUTES); // by default let optimizer wait a long time in case it retrieves some data from ConnectorMetadata
     private DataSize filterAndProjectMinOutputPageSize = DataSize.of(500, KILOBYTE);
     private int filterAndProjectMinOutputPageRowCount = 256;
     private int maxGroupingSets = 2048;
@@ -151,84 +104,12 @@ public class FeaturesConfig
 
     private boolean hideInaccessibleColumns;
 
-    public enum JoinReorderingStrategy
-    {
-        NONE,
-        ELIMINATE_CROSS_JOINS,
-        AUTOMATIC,
-    }
-
-    public enum JoinDistributionType
-    {
-        BROADCAST,
-        PARTITIONED,
-        AUTOMATIC;
-
-        public boolean canPartition()
-        {
-            return this == PARTITIONED || this == AUTOMATIC;
-        }
-
-        public boolean canReplicate()
-        {
-            return this == BROADCAST || this == AUTOMATIC;
-        }
-    }
-
     public enum DataIntegrityVerification
     {
         NONE,
         ABORT,
         RETRY,
         /**/;
-    }
-
-    public double getCpuCostWeight()
-    {
-        return cpuCostWeight;
-    }
-
-    @Config("cpu-cost-weight")
-    public FeaturesConfig setCpuCostWeight(double cpuCostWeight)
-    {
-        this.cpuCostWeight = cpuCostWeight;
-        return this;
-    }
-
-    public double getMemoryCostWeight()
-    {
-        return memoryCostWeight;
-    }
-
-    @Config("memory-cost-weight")
-    public FeaturesConfig setMemoryCostWeight(double memoryCostWeight)
-    {
-        this.memoryCostWeight = memoryCostWeight;
-        return this;
-    }
-
-    public double getNetworkCostWeight()
-    {
-        return networkCostWeight;
-    }
-
-    @Config("network-cost-weight")
-    public FeaturesConfig setNetworkCostWeight(double networkCostWeight)
-    {
-        this.networkCostWeight = networkCostWeight;
-        return this;
-    }
-
-    public boolean isDistributedIndexJoinsEnabled()
-    {
-        return distributedIndexJoinsEnabled;
-    }
-
-    @Config("distributed-index-joins-enabled")
-    public FeaturesConfig setDistributedIndexJoinsEnabled(boolean distributedIndexJoinsEnabled)
-    {
-        this.distributedIndexJoinsEnabled = distributedIndexJoinsEnabled;
-        return this;
     }
 
     public boolean isOmitDateTimeTypePrecision()
@@ -253,32 +134,6 @@ public class FeaturesConfig
     public FeaturesConfig setLegacyRowToJsonCast(boolean legacyRowToJsonCast)
     {
         this.legacyRowToJsonCast = legacyRowToJsonCast;
-        return this;
-    }
-
-    public JoinDistributionType getJoinDistributionType()
-    {
-        return joinDistributionType;
-    }
-
-    @Config("join-distribution-type")
-    public FeaturesConfig setJoinDistributionType(JoinDistributionType joinDistributionType)
-    {
-        this.joinDistributionType = requireNonNull(joinDistributionType, "joinDistributionType is null");
-        return this;
-    }
-
-    @NotNull
-    public DataSize getJoinMaxBroadcastTableSize()
-    {
-        return joinMaxBroadcastTableSize;
-    }
-
-    @Config("join-max-broadcast-table-size")
-    @ConfigDescription("Maximum estimated size of a table that can be broadcast when using automatic join type selection")
-    public FeaturesConfig setJoinMaxBroadcastTableSize(DataSize joinMaxBroadcastTableSize)
-    {
-        this.joinMaxBroadcastTableSize = joinMaxBroadcastTableSize;
         return this;
     }
 
@@ -323,59 +178,6 @@ public class FeaturesConfig
         return this;
     }
 
-    public boolean isColocatedJoinsEnabled()
-    {
-        return colocatedJoinsEnabled;
-    }
-
-    @Config("colocated-joins-enabled")
-    @ConfigDescription("Experimental: Use a colocated join when possible")
-    public FeaturesConfig setColocatedJoinsEnabled(boolean colocatedJoinsEnabled)
-    {
-        this.colocatedJoinsEnabled = colocatedJoinsEnabled;
-        return this;
-    }
-
-    public boolean isSpatialJoinsEnabled()
-    {
-        return spatialJoinsEnabled;
-    }
-
-    @Config("spatial-joins-enabled")
-    @ConfigDescription("Use spatial index for spatial joins when possible")
-    public FeaturesConfig setSpatialJoinsEnabled(boolean spatialJoinsEnabled)
-    {
-        this.spatialJoinsEnabled = spatialJoinsEnabled;
-        return this;
-    }
-
-    public JoinReorderingStrategy getJoinReorderingStrategy()
-    {
-        return joinReorderingStrategy;
-    }
-
-    @Config("optimizer.join-reordering-strategy")
-    @ConfigDescription("The strategy to use for reordering joins")
-    public FeaturesConfig setJoinReorderingStrategy(JoinReorderingStrategy joinReorderingStrategy)
-    {
-        this.joinReorderingStrategy = joinReorderingStrategy;
-        return this;
-    }
-
-    @Min(2)
-    public int getMaxReorderedJoins()
-    {
-        return maxReorderedJoins;
-    }
-
-    @Config("optimizer.max-reordered-joins")
-    @ConfigDescription("The maximum number of tables to reorder in cost-based join reordering")
-    public FeaturesConfig setMaxReorderedJoins(int maxReorderedJoins)
-    {
-        this.maxReorderedJoins = maxReorderedJoins;
-        return this;
-    }
-
     public boolean isRedistributeWrites()
     {
         return redistributeWrites;
@@ -385,32 +187,6 @@ public class FeaturesConfig
     public FeaturesConfig setRedistributeWrites(boolean redistributeWrites)
     {
         this.redistributeWrites = redistributeWrites;
-        return this;
-    }
-
-    public boolean isUsePreferredWritePartitioning()
-    {
-        return usePreferredWritePartitioning;
-    }
-
-    @Config("use-preferred-write-partitioning")
-    public FeaturesConfig setUsePreferredWritePartitioning(boolean usePreferredWritePartitioning)
-    {
-        this.usePreferredWritePartitioning = usePreferredWritePartitioning;
-        return this;
-    }
-
-    @Min(1)
-    public int getPreferredWritePartitioningMinNumberOfPartitions()
-    {
-        return preferredWritePartitioningMinNumberOfPartitions;
-    }
-
-    @Config("preferred-write-partitioning-min-number-of-partitions")
-    @ConfigDescription("Use preferred write partitioning when the number of written partitions exceeds the configured threshold")
-    public FeaturesConfig setPreferredWritePartitioningMinNumberOfPartitions(int preferredWritePartitioningMinNumberOfPartitions)
-    {
-        this.preferredWritePartitioningMinNumberOfPartitions = preferredWritePartitioningMinNumberOfPartitions;
         return this;
     }
 
@@ -437,91 +213,6 @@ public class FeaturesConfig
     public FeaturesConfig setWriterMinSize(DataSize writerMinSize)
     {
         this.writerMinSize = writerMinSize;
-        return this;
-    }
-
-    public boolean isOptimizeMetadataQueries()
-    {
-        return optimizeMetadataQueries;
-    }
-
-    @Config("optimizer.optimize-metadata-queries")
-    public FeaturesConfig setOptimizeMetadataQueries(boolean optimizeMetadataQueries)
-    {
-        this.optimizeMetadataQueries = optimizeMetadataQueries;
-        return this;
-    }
-
-    public boolean isUseMarkDistinct()
-    {
-        return useMarkDistinct;
-    }
-
-    @Config("optimizer.use-mark-distinct")
-    public FeaturesConfig setUseMarkDistinct(boolean value)
-    {
-        this.useMarkDistinct = value;
-        return this;
-    }
-
-    public boolean isPreferPartialAggregation()
-    {
-        return preferPartialAggregation;
-    }
-
-    @Config("optimizer.prefer-partial-aggregation")
-    public FeaturesConfig setPreferPartialAggregation(boolean value)
-    {
-        this.preferPartialAggregation = value;
-        return this;
-    }
-
-    public boolean isOptimizeTopNRanking()
-    {
-        return optimizeTopNRanking;
-    }
-
-    @Config("optimizer.optimize-top-n-ranking")
-    @LegacyConfig("optimizer.optimize-top-n-row-number")
-    public FeaturesConfig setOptimizeTopNRanking(boolean optimizeTopNRanking)
-    {
-        this.optimizeTopNRanking = optimizeTopNRanking;
-        return this;
-    }
-
-    public boolean isOptimizeHashGeneration()
-    {
-        return optimizeHashGeneration;
-    }
-
-    @Config("optimizer.optimize-hash-generation")
-    public FeaturesConfig setOptimizeHashGeneration(boolean optimizeHashGeneration)
-    {
-        this.optimizeHashGeneration = optimizeHashGeneration;
-        return this;
-    }
-
-    public boolean isPushTableWriteThroughUnion()
-    {
-        return pushTableWriteThroughUnion;
-    }
-
-    @Config("optimizer.push-table-write-through-union")
-    public FeaturesConfig setPushTableWriteThroughUnion(boolean pushTableWriteThroughUnion)
-    {
-        this.pushTableWriteThroughUnion = pushTableWriteThroughUnion;
-        return this;
-    }
-
-    public boolean isDictionaryAggregation()
-    {
-        return dictionaryAggregation;
-    }
-
-    @Config("optimizer.dictionary-aggregation")
-    public FeaturesConfig setDictionaryAggregation(boolean dictionaryAggregation)
-    {
-        this.dictionaryAggregation = dictionaryAggregation;
         return this;
     }
 
@@ -573,94 +264,6 @@ public class FeaturesConfig
     public FeaturesConfig setSpillEnabled(boolean spillEnabled)
     {
         this.spillEnabled = spillEnabled;
-        return this;
-    }
-
-    public Duration getIterativeOptimizerTimeout()
-    {
-        return iterativeOptimizerTimeout;
-    }
-
-    @Config("iterative-optimizer-timeout")
-    @LegacyConfig("experimental.iterative-optimizer-timeout")
-    public FeaturesConfig setIterativeOptimizerTimeout(Duration timeout)
-    {
-        this.iterativeOptimizerTimeout = timeout;
-        return this;
-    }
-
-    public boolean isEnableStatsCalculator()
-    {
-        return enableStatsCalculator;
-    }
-
-    @Config("enable-stats-calculator")
-    @LegacyConfig("experimental.enable-stats-calculator")
-    public FeaturesConfig setEnableStatsCalculator(boolean enableStatsCalculator)
-    {
-        this.enableStatsCalculator = enableStatsCalculator;
-        return this;
-    }
-
-    public boolean isStatisticsPrecalculationForPushdownEnabled()
-    {
-        return statisticsPrecalculationForPushdownEnabled;
-    }
-
-    @Config("statistics-precalculation-for-pushdown.enabled")
-    public FeaturesConfig setStatisticsPrecalculationForPushdownEnabled(boolean statisticsPrecalculationForPushdownEnabled)
-    {
-        this.statisticsPrecalculationForPushdownEnabled = statisticsPrecalculationForPushdownEnabled;
-        return this;
-    }
-
-    public boolean isCollectPlanStatisticsForAllQueries()
-    {
-        return collectPlanStatisticsForAllQueries;
-    }
-
-    @Config("collect-plan-statistics-for-all-queries")
-    @ConfigDescription("Collect plan statistics for non-EXPLAIN queries")
-    public FeaturesConfig setCollectPlanStatisticsForAllQueries(boolean collectPlanStatisticsForAllQueries)
-    {
-        this.collectPlanStatisticsForAllQueries = collectPlanStatisticsForAllQueries;
-        return this;
-    }
-
-    public boolean isIgnoreStatsCalculatorFailures()
-    {
-        return ignoreStatsCalculatorFailures;
-    }
-
-    @Config("optimizer.ignore-stats-calculator-failures")
-    @ConfigDescription("Ignore statistics calculator failures")
-    public FeaturesConfig setIgnoreStatsCalculatorFailures(boolean ignoreStatsCalculatorFailures)
-    {
-        this.ignoreStatsCalculatorFailures = ignoreStatsCalculatorFailures;
-        return this;
-    }
-
-    @Config("optimizer.default-filter-factor-enabled")
-    public FeaturesConfig setDefaultFilterFactorEnabled(boolean defaultFilterFactorEnabled)
-    {
-        this.defaultFilterFactorEnabled = defaultFilterFactorEnabled;
-        return this;
-    }
-
-    public boolean isDefaultFilterFactorEnabled()
-    {
-        return defaultFilterFactorEnabled;
-    }
-
-    public boolean isEnableForcedExchangeBelowGroupId()
-    {
-        return enableForcedExchangeBelowGroupId;
-    }
-
-    @Config("enable-forced-exchange-below-group-id")
-    public FeaturesConfig setEnableForcedExchangeBelowGroupId(boolean enableForcedExchangeBelowGroupId)
-    {
-        this.enableForcedExchangeBelowGroupId = enableForcedExchangeBelowGroupId;
         return this;
     }
 
@@ -750,18 +353,6 @@ public class FeaturesConfig
         return this;
     }
 
-    public boolean isOptimizeMixedDistinctAggregations()
-    {
-        return optimizeMixedDistinctAggregations;
-    }
-
-    @Config("optimizer.optimize-mixed-distinct-aggregations")
-    public FeaturesConfig setOptimizeMixedDistinctAggregations(boolean value)
-    {
-        this.optimizeMixedDistinctAggregations = value;
-        return this;
-    }
-
     public boolean isExchangeCompressionEnabled()
     {
         return exchangeCompressionEnabled;
@@ -786,43 +377,6 @@ public class FeaturesConfig
         return this;
     }
 
-    public boolean isEnableIntermediateAggregations()
-    {
-        return enableIntermediateAggregations;
-    }
-
-    @Config("optimizer.enable-intermediate-aggregations")
-    public FeaturesConfig setEnableIntermediateAggregations(boolean enableIntermediateAggregations)
-    {
-        this.enableIntermediateAggregations = enableIntermediateAggregations;
-        return this;
-    }
-
-    public boolean isPushAggregationThroughOuterJoin()
-    {
-        return pushAggregationThroughOuterJoin;
-    }
-
-    @Config("optimizer.push-aggregation-through-outer-join")
-    @LegacyConfig("optimizer.push-aggregation-through-join")
-    public FeaturesConfig setPushAggregationThroughOuterJoin(boolean pushAggregationThroughOuterJoin)
-    {
-        this.pushAggregationThroughOuterJoin = pushAggregationThroughOuterJoin;
-        return this;
-    }
-
-    public boolean isPushPartialAggregationThoughJoin()
-    {
-        return pushPartialAggregationThoughJoin;
-    }
-
-    @Config("optimizer.push-partial-aggregation-through-join")
-    public FeaturesConfig setPushPartialAggregationThoughJoin(boolean pushPartialAggregationThoughJoin)
-    {
-        this.pushPartialAggregationThoughJoin = pushPartialAggregationThoughJoin;
-        return this;
-    }
-
     public boolean isParseDecimalLiteralsAsDouble()
     {
         return parseDecimalLiteralsAsDouble;
@@ -832,18 +386,6 @@ public class FeaturesConfig
     public FeaturesConfig setParseDecimalLiteralsAsDouble(boolean parseDecimalLiteralsAsDouble)
     {
         this.parseDecimalLiteralsAsDouble = parseDecimalLiteralsAsDouble;
-        return this;
-    }
-
-    public boolean isForceSingleNodeOutput()
-    {
-        return forceSingleNodeOutput;
-    }
-
-    @Config("optimizer.force-single-node-output")
-    public FeaturesConfig setForceSingleNodeOutput(boolean value)
-    {
-        this.forceSingleNodeOutput = value;
         return this;
     }
 
@@ -887,18 +429,6 @@ public class FeaturesConfig
         return this;
     }
 
-    public boolean isDistributedSortEnabled()
-    {
-        return distributedSort;
-    }
-
-    @Config("distributed-sort")
-    public FeaturesConfig setDistributedSortEnabled(boolean enabled)
-    {
-        distributedSort = enabled;
-        return this;
-    }
-
     public int getMaxRecursionDepth()
     {
         return maxRecursionDepth;
@@ -934,130 +464,6 @@ public class FeaturesConfig
     public FeaturesConfig setLateMaterializationEnabled(boolean lateMaterializationEnabled)
     {
         this.lateMaterializationEnabled = lateMaterializationEnabled;
-        return this;
-    }
-
-    public boolean isSkipRedundantSort()
-    {
-        return skipRedundantSort;
-    }
-
-    @Config("optimizer.skip-redundant-sort")
-    public FeaturesConfig setSkipRedundantSort(boolean value)
-    {
-        this.skipRedundantSort = value;
-        return this;
-    }
-
-    public boolean isComplexExpressionPushdownEnabled()
-    {
-        return complexExpressionPushdownEnabled;
-    }
-
-    @Config("optimizer.complex-expression-pushdown.enabled")
-    public FeaturesConfig setComplexExpressionPushdownEnabled(boolean complexExpressionPushdownEnabled)
-    {
-        this.complexExpressionPushdownEnabled = complexExpressionPushdownEnabled;
-        return this;
-    }
-
-    public boolean isPredicatePushdownUseTableProperties()
-    {
-        return predicatePushdownUseTableProperties;
-    }
-
-    @Config("optimizer.predicate-pushdown-use-table-properties")
-    public FeaturesConfig setPredicatePushdownUseTableProperties(boolean predicatePushdownUseTableProperties)
-    {
-        this.predicatePushdownUseTableProperties = predicatePushdownUseTableProperties;
-        return this;
-    }
-
-    public boolean isIgnoreDownstreamPreferences()
-    {
-        return ignoreDownstreamPreferences;
-    }
-
-    @Config("optimizer.ignore-downstream-preferences")
-    public FeaturesConfig setIgnoreDownstreamPreferences(boolean ignoreDownstreamPreferences)
-    {
-        this.ignoreDownstreamPreferences = ignoreDownstreamPreferences;
-        return this;
-    }
-
-    public boolean isRewriteFilteringSemiJoinToInnerJoin()
-    {
-        return rewriteFilteringSemiJoinToInnerJoin;
-    }
-
-    @Config("optimizer.rewrite-filtering-semi-join-to-inner-join")
-    public FeaturesConfig setRewriteFilteringSemiJoinToInnerJoin(boolean rewriteFilteringSemiJoinToInnerJoin)
-    {
-        this.rewriteFilteringSemiJoinToInnerJoin = rewriteFilteringSemiJoinToInnerJoin;
-        return this;
-    }
-
-    public boolean isOptimizeDuplicateInsensitiveJoins()
-    {
-        return optimizeDuplicateInsensitiveJoins;
-    }
-
-    @Config("optimizer.optimize-duplicate-insensitive-joins")
-    public FeaturesConfig setOptimizeDuplicateInsensitiveJoins(boolean optimizeDuplicateInsensitiveJoins)
-    {
-        this.optimizeDuplicateInsensitiveJoins = optimizeDuplicateInsensitiveJoins;
-        return this;
-    }
-
-    public boolean isUseLegacyWindowFilterPushdown()
-    {
-        return useLegacyWindowFilterPushdown;
-    }
-
-    @Config("optimizer.use-legacy-window-filter-pushdown")
-    public FeaturesConfig setUseLegacyWindowFilterPushdown(boolean useLegacyWindowFilterPushdown)
-    {
-        this.useLegacyWindowFilterPushdown = useLegacyWindowFilterPushdown;
-        return this;
-    }
-
-    public boolean isUseTableScanNodePartitioning()
-    {
-        return useTableScanNodePartitioning;
-    }
-
-    @Config("optimizer.use-table-scan-node-partitioning")
-    @LegacyConfig("optimizer.plan-with-table-node-partitioning")
-    @ConfigDescription("Adapt plan to node pre-partitioned tables")
-    public FeaturesConfig setUseTableScanNodePartitioning(boolean useTableScanNodePartitioning)
-    {
-        this.useTableScanNodePartitioning = useTableScanNodePartitioning;
-        return this;
-    }
-
-    @Min(0)
-    public double getTableScanNodePartitioningMinBucketToTaskRatio()
-    {
-        return tableScanNodePartitioningMinBucketToTaskRatio;
-    }
-
-    @Config("optimizer.table-scan-node-partitioning-min-bucket-to-task-ratio")
-    @ConfigDescription("Min table scan bucket to task ratio for which plan will be adopted to node pre-partitioned tables")
-    public FeaturesConfig setTableScanNodePartitioningMinBucketToTaskRatio(double tableScanNodePartitioningMinBucketToTaskRatio)
-    {
-        this.tableScanNodePartitioningMinBucketToTaskRatio = tableScanNodePartitioningMinBucketToTaskRatio;
-        return this;
-    }
-
-    public boolean isMergeProjectWithValues()
-    {
-        return mergeProjectWithValues;
-    }
-
-    @Config("optimizer.merge-project-with-values")
-    public FeaturesConfig setMergeProjectWithValues(boolean mergeProjectWithValues)
-    {
-        this.mergeProjectWithValues = mergeProjectWithValues;
         return this;
     }
