+++ core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/EliminateCrossJoins.java
@@ -16,8 +16,8 @@ package io.trino.sql.planner.iterative.rule;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
-import io.trino.FeaturesConfig.JoinReorderingStrategy;
 import io.trino.Session;
+import io.trino.cost.OptimizerConfig.JoinReorderingStrategy;
 import io.trino.matching.Captures;
 import io.trino.matching.Pattern;
 import io.trino.sql.PlannerContext;
@@ -43,9 +43,9 @@ import java.util.Set;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
 import static com.google.common.collect.ImmutableList.toImmutableList;
-import static io.trino.FeaturesConfig.JoinReorderingStrategy.AUTOMATIC;
-import static io.trino.FeaturesConfig.JoinReorderingStrategy.ELIMINATE_CROSS_JOINS;
 import static io.trino.SystemSessionProperties.getJoinReorderingStrategy;
+import static io.trino.cost.OptimizerConfig.JoinReorderingStrategy.AUTOMATIC;
+import static io.trino.cost.OptimizerConfig.JoinReorderingStrategy.ELIMINATE_CROSS_JOINS;
 import static io.trino.sql.planner.iterative.rule.Util.restrictOutputs;
 import static io.trino.sql.planner.plan.Patterns.join;
 import static java.util.Comparator.comparing;
