+++ core/trino-main/src/main/java/io/trino/testing/LocalQueryRunner.java
@@ -37,6 +37,7 @@ import io.trino.connector.system.SchemaPropertiesSystemTable;
 import io.trino.connector.system.TableCommentSystemTable;
 import io.trino.connector.system.TablePropertiesSystemTable;
 import io.trino.connector.system.TransactionsSystemTable;
+import io.trino.cost.OptimizerConfig;
 import io.trino.cost.ComposableStatsCalculator;
 import io.trino.cost.CostCalculator;
 import io.trino.cost.CostCalculatorUsingExchanges;
@@ -271,7 +272,7 @@ public class LocalQueryRunner
     private final TaskManagerConfig taskManagerConfig;
     private final boolean alwaysRevokeMemory;
     private final NodeSpillConfig nodeSpillConfig;
-    private final FeaturesConfig featuresConfig;
+    private final OptimizerConfig optimizerConfig;
     private final PlanOptimizersProvider planOptimizersProvider;
     private final OperatorFactories operatorFactories;
     private final StatementAnalyzerFactory statementAnalyzerFactory;
@@ -324,7 +325,8 @@ public class LocalQueryRunner
         this.indexManager = new IndexManager();
         NodeSchedulerConfig nodeSchedulerConfig = new NodeSchedulerConfig().setIncludeCoordinator(true);
         NodeScheduler nodeScheduler = new NodeScheduler(new UniformNodeSelectorFactory(nodeManager, nodeSchedulerConfig, new NodeTaskMap(finalizerService)));
-        this.featuresConfig = requireNonNull(featuresConfig, "featuresConfig is null");
+        requireNonNull(featuresConfig, "featuresConfig is null");
+        this.optimizerConfig = new OptimizerConfig();
         this.pageSinkManager = new PageSinkManager();
         this.catalogManager = new CatalogManager();
         this.transactionManager = InMemoryTransactionManager.create(
@@ -357,7 +359,7 @@ public class LocalQueryRunner
         accessControl.loadSystemAccessControl(AllowAllSystemAccessControl.NAME, ImmutableMap.of());
 
         TableProceduresRegistry tableProceduresRegistry = new TableProceduresRegistry();
-        this.sessionPropertyManager = createSessionPropertyManager(extraSessionProperties, taskManagerConfig, featuresConfig);
+        this.sessionPropertyManager = createSessionPropertyManager(extraSessionProperties, taskManagerConfig, featuresConfig, optimizerConfig);
         this.schemaPropertyManager = new SchemaPropertyManager();
         this.columnPropertyManager = new ColumnPropertyManager();
         this.tablePropertyManager = new TablePropertyManager();
@@ -489,7 +491,8 @@ public class LocalQueryRunner
     private static SessionPropertyManager createSessionPropertyManager(
             Set<SystemSessionPropertiesProvider> extraSessionProperties,
             TaskManagerConfig taskManagerConfig,
-            FeaturesConfig featuresConfig)
+            FeaturesConfig featuresConfig,
+            OptimizerConfig optimizerConfig)
     {
         Set<SystemSessionPropertiesProvider> systemSessionProperties = ImmutableSet.<SystemSessionPropertiesProvider>builder()
                 .addAll(requireNonNull(extraSessionProperties, "extraSessionProperties is null"))
@@ -498,6 +501,7 @@ public class LocalQueryRunner
                         taskManagerConfig,
                         new MemoryManagerConfig(),
                         featuresConfig,
+                        optimizerConfig,
                         new NodeMemoryConfig(),
                         new DynamicFilterConfig(),
                         new NodeSchedulerConfig()))
@@ -1007,7 +1011,7 @@ public class LocalQueryRunner
                 scalarStatsCalculator,
                 costCalculator,
                 estimatedExchangesCostCalculator,
-                new CostComparator(featuresConfig),
+                new CostComparator(optimizerConfig),
                 taskCountEstimator,
                 nodePartitioningManager,
                 new RuleStatsRecorder()).get();
