+++ core/trino-main/src/test/java/io/trino/sql/analyzer/TestFeaturesConfig.java
@@ -15,11 +15,8 @@ package io.trino.sql.analyzer;
 
 import com.google.common.collect.ImmutableMap;
 import io.airlift.units.DataSize;
-import io.airlift.units.Duration;
 import io.trino.FeaturesConfig;
 import io.trino.FeaturesConfig.DataIntegrityVerification;
-import io.trino.FeaturesConfig.JoinDistributionType;
-import io.trino.FeaturesConfig.JoinReorderingStrategy;
 import org.testng.annotations.Test;
 
 import java.util.Map;
@@ -30,12 +27,8 @@ import static io.airlift.configuration.testing.ConfigAssertions.recordDefaults;
 import static io.airlift.units.DataSize.Unit.GIGABYTE;
 import static io.airlift.units.DataSize.Unit.KILOBYTE;
 import static io.airlift.units.DataSize.Unit.MEGABYTE;
-import static io.trino.FeaturesConfig.JoinDistributionType.BROADCAST;
-import static io.trino.FeaturesConfig.JoinReorderingStrategy.NONE;
 import static io.trino.sql.analyzer.RegexLibrary.JONI;
 import static io.trino.sql.analyzer.RegexLibrary.RE2J;
-import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.concurrent.TimeUnit.SECONDS;
 
 public class TestFeaturesConfig
 {
@@ -43,28 +36,12 @@ public class TestFeaturesConfig
     public void testDefaults()
     {
         assertRecordedDefaults(recordDefaults(FeaturesConfig.class)
-                .setCpuCostWeight(75)
-                .setMemoryCostWeight(10)
-                .setNetworkCostWeight(15)
-                .setDistributedIndexJoinsEnabled(false)
-                .setJoinMaxBroadcastTableSize(DataSize.of(100, MEGABYTE))
-                .setJoinDistributionType(JoinDistributionType.AUTOMATIC)
                 .setGroupedExecutionEnabled(false)
                 .setDynamicScheduleForGroupedExecutionEnabled(false)
                 .setConcurrentLifespansPerTask(0)
-                .setColocatedJoinsEnabled(false)
-                .setSpatialJoinsEnabled(true)
-                .setJoinReorderingStrategy(JoinReorderingStrategy.AUTOMATIC)
-                .setMaxReorderedJoins(9)
                 .setRedistributeWrites(true)
-                .setUsePreferredWritePartitioning(true)
-                .setPreferredWritePartitioningMinNumberOfPartitions(50)
                 .setScaleWriters(false)
                 .setWriterMinSize(DataSize.of(32, MEGABYTE))
-                .setOptimizeMetadataQueries(false)
-                .setOptimizeHashGeneration(true)
-                .setPushTableWriteThroughUnion(true)
-                .setDictionaryAggregation(false)
                 .setRegexLibrary(JONI)
                 .setRe2JDfaStatesLimit(Integer.MAX_VALUE)
                 .setRe2JDfaRetries(5)
@@ -75,43 +52,17 @@ public class TestFeaturesConfig
                 .setSpillMaxUsedSpaceThreshold(0.9)
                 .setMemoryRevokingThreshold(0.9)
                 .setMemoryRevokingTarget(0.5)
-                .setOptimizeMixedDistinctAggregations(false)
-                .setIterativeOptimizerTimeout(new Duration(3, MINUTES))
-                .setEnableStatsCalculator(true)
-                .setStatisticsPrecalculationForPushdownEnabled(false)
-                .setCollectPlanStatisticsForAllQueries(false)
-                .setIgnoreStatsCalculatorFailures(true)
-                .setDefaultFilterFactorEnabled(false)
-                .setEnableForcedExchangeBelowGroupId(true)
                 .setExchangeCompressionEnabled(false)
                 .setExchangeDataIntegrityVerification(DataIntegrityVerification.ABORT)
                 .setLegacyRowToJsonCast(false)
-                .setEnableIntermediateAggregations(false)
-                .setPushAggregationThroughOuterJoin(true)
-                .setPushPartialAggregationThoughJoin(false)
                 .setParseDecimalLiteralsAsDouble(false)
-                .setForceSingleNodeOutput(true)
                 .setPagesIndexEagerCompactionEnabled(false)
                 .setFilterAndProjectMinOutputPageSize(DataSize.of(500, KILOBYTE))
                 .setFilterAndProjectMinOutputPageRowCount(256)
-                .setUseMarkDistinct(true)
-                .setPreferPartialAggregation(true)
-                .setOptimizeTopNRanking(true)
-                .setDistributedSortEnabled(true)
                 .setMaxRecursionDepth(10)
                 .setMaxGroupingSets(2048)
                 .setLateMaterializationEnabled(false)
-                .setSkipRedundantSort(true)
-                .setComplexExpressionPushdownEnabled(true)
-                .setPredicatePushdownUseTableProperties(true)
-                .setIgnoreDownstreamPreferences(false)
                 .setOmitDateTimeTypePrecision(false)
-                .setRewriteFilteringSemiJoinToInnerJoin(true)
-                .setOptimizeDuplicateInsensitiveJoins(true)
-                .setUseLegacyWindowFilterPushdown(false)
-                .setUseTableScanNodePartitioning(true)
-                .setTableScanNodePartitioningMinBucketToTaskRatio(0.5)
-                .setMergeProjectWithValues(true)
                 .setLegacyCatalogRoles(false)
                 .setIncrementalHashArrayLoadFactorEnabled(true)
                 .setHideInaccessibleColumns(false)
@@ -122,38 +73,12 @@ public class TestFeaturesConfig
     public void testExplicitPropertyMappings()
     {
         Map<String, String> properties = ImmutableMap.<String, String>builder()
-                .put("cpu-cost-weight", "0.4")
-                .put("memory-cost-weight", "0.3")
-                .put("network-cost-weight", "0.2")
-                .put("iterative-optimizer-timeout", "10s")
-                .put("enable-stats-calculator", "false")
-                .put("statistics-precalculation-for-pushdown.enabled", "true")
-                .put("collect-plan-statistics-for-all-queries", "true")
-                .put("optimizer.ignore-stats-calculator-failures", "false")
-                .put("optimizer.default-filter-factor-enabled", "true")
-                .put("enable-forced-exchange-below-group-id", "false")
-                .put("distributed-index-joins-enabled", "true")
-                .put("join-distribution-type", "BROADCAST")
-                .put("join-max-broadcast-table-size", "42GB")
                 .put("grouped-execution-enabled", "true")
                 .put("dynamic-schedule-for-grouped-execution", "true")
                 .put("concurrent-lifespans-per-task", "1")
-                .put("colocated-joins-enabled", "true")
-                .put("spatial-joins-enabled", "false")
-                .put("optimizer.join-reordering-strategy", "NONE")
-                .put("optimizer.max-reordered-joins", "5")
                 .put("redistribute-writes", "false")
-                .put("use-preferred-write-partitioning", "false")
-                .put("preferred-write-partitioning-min-number-of-partitions", "10")
                 .put("scale-writers", "true")
                 .put("writer-min-size", "42GB")
-                .put("optimizer.optimize-metadata-queries", "true")
-                .put("optimizer.optimize-hash-generation", "false")
-                .put("optimizer.optimize-mixed-distinct-aggregations", "true")
-                .put("optimizer.push-table-write-through-union", "false")
-                .put("optimizer.dictionary-aggregation", "true")
-                .put("optimizer.push-aggregation-through-outer-join", "false")
-                .put("optimizer.push-partial-aggregation-through-join", "true")
                 .put("regex-library", "RE2J")
                 .put("re2j.dfa-states-limit", "42")
                 .put("re2j.dfa-retries", "42")
@@ -167,30 +92,14 @@ public class TestFeaturesConfig
                 .put("exchange.compression-enabled", "true")
                 .put("exchange.data-integrity-verification", "RETRY")
                 .put("deprecated.legacy-row-to-json-cast", "true")
-                .put("optimizer.enable-intermediate-aggregations", "true")
                 .put("parse-decimal-literals-as-double", "true")
-                .put("optimizer.force-single-node-output", "false")
                 .put("pages-index.eager-compaction-enabled", "true")
                 .put("filter-and-project-min-output-page-size", "1MB")
                 .put("filter-and-project-min-output-page-row-count", "2048")
-                .put("optimizer.use-mark-distinct", "false")
-                .put("optimizer.prefer-partial-aggregation", "false")
-                .put("optimizer.optimize-top-n-ranking", "false")
-                .put("distributed-sort", "false")
                 .put("max-recursion-depth", "8")
                 .put("analyzer.max-grouping-sets", "2047")
                 .put("experimental.late-materialization.enabled", "true")
-                .put("optimizer.skip-redundant-sort", "false")
-                .put("optimizer.complex-expression-pushdown.enabled", "false")
-                .put("optimizer.predicate-pushdown-use-table-properties", "false")
-                .put("optimizer.ignore-downstream-preferences", "true")
                 .put("deprecated.omit-datetime-type-precision", "true")
-                .put("optimizer.rewrite-filtering-semi-join-to-inner-join", "false")
-                .put("optimizer.optimize-duplicate-insensitive-joins", "false")
-                .put("optimizer.use-legacy-window-filter-pushdown", "true")
-                .put("optimizer.use-table-scan-node-partitioning", "false")
-                .put("optimizer.table-scan-node-partitioning-min-bucket-to-task-ratio", "0.0")
-                .put("optimizer.merge-project-with-values", "false")
                 .put("deprecated.legacy-catalog-roles", "true")
                 .put("incremental-hash-array-load-factor.enabled", "false")
                 .put("hide-inaccessible-columns", "true")
@@ -198,37 +107,12 @@ public class TestFeaturesConfig
                 .buildOrThrow();
 
         FeaturesConfig expected = new FeaturesConfig()
-                .setCpuCostWeight(0.4)
-                .setMemoryCostWeight(0.3)
-                .setNetworkCostWeight(0.2)
-                .setIterativeOptimizerTimeout(new Duration(10, SECONDS))
-                .setEnableStatsCalculator(false)
-                .setStatisticsPrecalculationForPushdownEnabled(true)
-                .setCollectPlanStatisticsForAllQueries(true)
-                .setIgnoreStatsCalculatorFailures(false)
-                .setEnableForcedExchangeBelowGroupId(false)
-                .setDistributedIndexJoinsEnabled(true)
-                .setJoinDistributionType(BROADCAST)
-                .setJoinMaxBroadcastTableSize(DataSize.of(42, GIGABYTE))
                 .setGroupedExecutionEnabled(true)
                 .setDynamicScheduleForGroupedExecutionEnabled(true)
                 .setConcurrentLifespansPerTask(1)
-                .setColocatedJoinsEnabled(true)
-                .setSpatialJoinsEnabled(false)
-                .setJoinReorderingStrategy(NONE)
-                .setMaxReorderedJoins(5)
                 .setRedistributeWrites(false)
-                .setUsePreferredWritePartitioning(false)
-                .setPreferredWritePartitioningMinNumberOfPartitions(10)
                 .setScaleWriters(true)
                 .setWriterMinSize(DataSize.of(42, GIGABYTE))
-                .setOptimizeMetadataQueries(true)
-                .setOptimizeHashGeneration(false)
-                .setOptimizeMixedDistinctAggregations(true)
-                .setPushTableWriteThroughUnion(false)
-                .setDictionaryAggregation(true)
-                .setPushAggregationThroughOuterJoin(false)
-                .setPushPartialAggregationThoughJoin(true)
                 .setRegexLibrary(RE2J)
                 .setRe2JDfaStatesLimit(42)
                 .setRe2JDfaRetries(42)
@@ -242,31 +126,14 @@ public class TestFeaturesConfig
                 .setExchangeCompressionEnabled(true)
                 .setExchangeDataIntegrityVerification(DataIntegrityVerification.RETRY)
                 .setLegacyRowToJsonCast(true)
-                .setEnableIntermediateAggregations(true)
                 .setParseDecimalLiteralsAsDouble(true)
-                .setForceSingleNodeOutput(false)
                 .setPagesIndexEagerCompactionEnabled(true)
                 .setFilterAndProjectMinOutputPageSize(DataSize.of(1, MEGABYTE))
                 .setFilterAndProjectMinOutputPageRowCount(2048)
-                .setUseMarkDistinct(false)
-                .setPreferPartialAggregation(false)
-                .setOptimizeTopNRanking(false)
-                .setDistributedSortEnabled(false)
                 .setMaxRecursionDepth(8)
                 .setMaxGroupingSets(2047)
-                .setDefaultFilterFactorEnabled(true)
                 .setLateMaterializationEnabled(true)
-                .setSkipRedundantSort(false)
-                .setComplexExpressionPushdownEnabled(false)
-                .setPredicatePushdownUseTableProperties(false)
-                .setIgnoreDownstreamPreferences(true)
                 .setOmitDateTimeTypePrecision(true)
-                .setRewriteFilteringSemiJoinToInnerJoin(false)
-                .setOptimizeDuplicateInsensitiveJoins(false)
-                .setUseLegacyWindowFilterPushdown(true)
-                .setUseTableScanNodePartitioning(false)
-                .setTableScanNodePartitioningMinBucketToTaskRatio(0.0)
-                .setMergeProjectWithValues(false)
                 .setLegacyCatalogRoles(true)
                 .setIncrementalHashArrayLoadFactorEnabled(false)
                 .setHideInaccessibleColumns(true)
