+++ core/trino-main/src/test/java/io/trino/sql/planner/TestLocalDynamicFilterConsumer.java
@@ -18,8 +18,8 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.util.concurrent.ListenableFuture;
-import io.trino.FeaturesConfig.JoinDistributionType;
-import io.trino.FeaturesConfig.JoinReorderingStrategy;
+import io.trino.cost.OptimizerConfig.JoinDistributionType;
+import io.trino.cost.OptimizerConfig.JoinReorderingStrategy;
 import io.trino.spi.predicate.Domain;
 import io.trino.spi.predicate.TupleDomain;
 import io.trino.sql.planner.assertions.BasePlanTest;
