+++ core/trino-main/src/test/java/io/trino/sql/planner/optimizations/TestUnionWithReplicatedJoin.java
@@ -15,8 +15,8 @@ package io.trino.sql.planner.optimizations;
 
 import com.google.common.collect.ImmutableMap;
 
-import static io.trino.FeaturesConfig.JoinDistributionType.BROADCAST;
 import static io.trino.SystemSessionProperties.JOIN_DISTRIBUTION_TYPE;
+import static io.trino.cost.OptimizerConfig.JoinDistributionType.BROADCAST;
 
 public class TestUnionWithReplicatedJoin
         extends TestUnion
