+++ core/trino-main/src/test/java/io/trino/sql/query/TestFilterHideInacessibleColumnsSession.java
@@ -15,6 +15,7 @@ package io.trino.sql.query;
 
 import io.trino.FeaturesConfig;
 import io.trino.SystemSessionProperties;
+import io.trino.cost.OptimizerConfig;
 import io.trino.execution.DynamicFilterConfig;
 import io.trino.execution.QueryManagerConfig;
 import io.trino.execution.TaskManagerConfig;
@@ -39,6 +40,7 @@ public class TestFilterHideInacessibleColumnsSession
                 new TaskManagerConfig(),
                 new MemoryManagerConfig(),
                 featuresConfig,
+                new OptimizerConfig(),
                 new NodeMemoryConfig(),
                 new DynamicFilterConfig(),
                 new NodeSchedulerConfig()));
@@ -56,6 +58,7 @@ public class TestFilterHideInacessibleColumnsSession
                 new TaskManagerConfig(),
                 new MemoryManagerConfig(),
                 featuresConfig,
+                new OptimizerConfig(),
                 new NodeMemoryConfig(),
                 new DynamicFilterConfig(),
                 new NodeSchedulerConfig()));
@@ -71,6 +74,7 @@ public class TestFilterHideInacessibleColumnsSession
                 new TaskManagerConfig(),
                 new MemoryManagerConfig(),
                 featuresConfig,
+                new OptimizerConfig(),
                 new NodeMemoryConfig(),
                 new DynamicFilterConfig(),
                 new NodeSchedulerConfig()));
@@ -86,6 +90,7 @@ public class TestFilterHideInacessibleColumnsSession
                 new TaskManagerConfig(),
                 new MemoryManagerConfig(),
                 featuresConfig,
+                new OptimizerConfig(),
                 new NodeMemoryConfig(),
                 new DynamicFilterConfig(),
                 new NodeSchedulerConfig()));
