+++ testing/trino-benchto-benchmarks/src/test/java/io/trino/sql/planner/TestTpcdsCostBasedPlan.java
@@ -15,9 +15,9 @@
 package io.trino.sql.planner;
 
 import com.google.common.collect.ImmutableMap;
-import io.trino.FeaturesConfig.JoinDistributionType;
-import io.trino.FeaturesConfig.JoinReorderingStrategy;
 import io.trino.Session;
+import io.trino.cost.OptimizerConfig.JoinDistributionType;
+import io.trino.cost.OptimizerConfig.JoinReorderingStrategy;
 import io.trino.plugin.tpcds.TpcdsConnectorFactory;
 import io.trino.testing.LocalQueryRunner;
