+++ testing/trino-testing/src/main/java/io/trino/testing/AbstractTestIntegrationSmokeTest.java
@@ -13,8 +13,8 @@
  */
 package io.trino.testing;
 
-import io.trino.FeaturesConfig.JoinDistributionType;
 import io.trino.Session;
+import io.trino.cost.OptimizerConfig.JoinDistributionType;
 import org.intellij.lang.annotations.Language;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
