+++ testing/trino-tests/src/test/java/io/trino/execution/TestCoordinatorDynamicFiltering.java
@@ -61,10 +61,6 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.stream.LongStream;
 
 import static com.google.common.collect.ImmutableList.toImmutableList;
-import static io.trino.FeaturesConfig.JoinDistributionType;
-import static io.trino.FeaturesConfig.JoinDistributionType.BROADCAST;
-import static io.trino.FeaturesConfig.JoinDistributionType.PARTITIONED;
-import static io.trino.FeaturesConfig.JoinReorderingStrategy.NONE;
 import static io.trino.SystemSessionProperties.ENABLE_COORDINATOR_DYNAMIC_FILTERS_DISTRIBUTION;
 import static io.trino.SystemSessionProperties.FILTERING_SEMI_JOIN_TO_INNER;
 import static io.trino.SystemSessionProperties.JOIN_DISTRIBUTION_TYPE;
@@ -72,6 +68,10 @@ import static io.trino.SystemSessionProperties.JOIN_REORDERING_STRATEGY;
 import static io.trino.SystemSessionProperties.TASK_CONCURRENCY;
 import static io.trino.SystemSessionProperties.getJoinDistributionType;
 import static io.trino.SystemSessionProperties.isEnableCoordinatorDynamicFiltersDistribution;
+import static io.trino.cost.OptimizerConfig.JoinDistributionType;
+import static io.trino.cost.OptimizerConfig.JoinDistributionType.BROADCAST;
+import static io.trino.cost.OptimizerConfig.JoinDistributionType.PARTITIONED;
+import static io.trino.cost.OptimizerConfig.JoinReorderingStrategy.NONE;
 import static io.trino.spi.predicate.Domain.multipleValues;
 import static io.trino.spi.predicate.Domain.singleValue;
 import static io.trino.spi.predicate.Range.range;
